Add `implements(_)` on Class
Provide a way to determine at runtime of an object has a given
method by letting you ask its class if it implements a method
with the given signature.

The decision to put this on Class rather than on Object (in other
words, making you ask the class rather than the instance) is
mostly driven by the lack of a good way to implement
`super.implements`

Even in this design, asking a class if it implements a method with
the given signature returns true whether or not the implementation
came from the class, or any if its superclasses. Because methods
are copied into classes, I don't think there's any way at runtime
to determine more information than this.